+++ app/src/main/java/io/gitlab/fsc_clam/fscwhereswhat/repo/impl/FakeRamCentralRepository.kt
@@ -27,9 +27,9 @@ import java.net.URL
 class FakeRamCentralRepository : RamCentralRepository {
 
 	//Internal storage for events
-	private val events: MutableMap<Int, Event> = mutableMapOf(
-		0 to Event(
-			0,
+	private val events: MutableMap<Long, Event> = mutableMapOf(
+		0L to Event(
+			0L,
 			"Basketball",
 			URL("https://tse1.mm.bing.net/th?id=OIP.OZQP0Ud2cFFmyo6yphrd1QAAAA&pid=Api"),
 			"Basketball Competition",
@@ -48,7 +48,7 @@ class FakeRamCentralRepository : RamCentralRepository {
 	 * @return The event with the given ID.
 	 * @throws NoSuchElementException if no event with the specified ID is found.
 	 */
-	override suspend fun getEvent(id: Int): Event {
+	override suspend fun getEvent(id: Long): Event {
 		return events[id] ?: throw NoSuchElementException("Event with id $id not found")
 	}
